requirements.txt
PyPDF2
pdf2image
Pillow
pyzbar
python-dotenv
loguru
tqdm
pytesseract
PySide6
numpy
opencv-python